Youtube Video Summary/requirements.txt
flask==2.3.3
openai==1.3.0
yt-dlp==2023.7.6
python-dotenv==1.0.0
flask-limiter==3.5.0
flask-talisman==1.0.0
pyyaml==6.0.1
